Add basic virtual camera functionality
Implemented an option to use a pre-recorded video as input instead of the physical camera.
This allows the program to run and be tested without needing to connect to the real camera at the hockey table.
Change the URL in constants.py to "virtual" to use the virtual camera.
The current video might not be perfect (slightly wrong fps and compressed) but should be good enough for now.